requirements.txt
pdfplumber==0.10.3
pytesseract==0.3.10
pillow==10.0.1
opencv-python==4.8.1.78
spacy==3.7.2
openai==1.3.7
pdf2image==1.16.3  
requests==2.31.0
pytest==7.4.3  
pyyaml==6.0.1  
python-dotenv==1.0.0 
presidio-analyzer==2.2.359  
pytest-mock==3.12.0